requirements.txt
Flask
python-dotenv
openai
chromadb
Flask-Cors
gunicorn